+++ packages/account-connector/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sophon-labs/account-connector
 
+## 3.0.0-next.36
+
+### Patch Changes
+
+- React native support
+- Updated dependencies
+  - @sophon-labs/account-provider@3.0.0-next.36
+  - @sophon-labs/account-core@3.0.0-next.36
+
 ## 3.0.0-next.35
 
 ### Patch Changes