+++ packages/account-partner/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sophon-labs/account-partner
 
+## 3.0.0-next.36
+
+### Patch Changes
+
+- React native support
+- Updated dependencies
+  - @sophon-labs/account-core@3.0.0-next.36
+
 ## 3.0.0-next.35
 
 ### Patch Changes